% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise_plan.R
\name{exercise_plan}
\alias{exercise_plan}
\title{Create an exercise plan based on current weight, height, age and target BMI.}
\usage{
exercise_plan(
  weight,
  height,
  sex,
  age,
  target_weight,
  number_of_days,
  return_graph = FALSE
)
}
\arguments{
\item{weight}{A numeric vector of body weight value, in kilograms.}

\item{height}{A numeric vector of height, in meters.}

\item{sex}{An integer vector of sex, used in the Harris–Benedict equation
estimation, 1 for male and 2 for female.}

\item{age}{An integer vector of current age, in years.}

\item{target_weight}{A numeric vector of target weight, in kilograms.}

\item{number_of_days}{An integer vector of number of days to reach the
target weight.}

\item{return_graph}{TRUE or FALSE on whether to return a graphical
visualization instead of a dictionary.}
}
\value{
A dictionary of recommended exercises, or a graph plotting that
}
\description{
The calorie consumption needed each day is computed and a set of activities
is suggested accordingly. If \code{return_graph} is \code{True}, a \code{ggplot} object
instead of a dictionary.
}
\examples{
exercise_plan(100, 1.83, 2, 27, 68, 30)
}
