% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_bmi.R
\name{project_bmi}
\alias{project_bmi}
\title{Compute average BMI change per week based on current weight, height, age,
and target BMI. The average BMI increase or decrease each week is computed.
If \code{return_graph}is \code{TRUE}, a \code{ggplot} graph is given based on forcasted
BMI trajectory for the target timeframe.}
\usage{
project_bmi(weight, height, target_bmi, number_of_days, return_graph = FALSE)
}
\arguments{
\item{weight}{A positive numeric value of body weight value in kilograms.}

\item{height}{A positive numeric value of height value in meters.}

\item{target_bmi}{A positive numeric value of target BMI}

\item{number_of_days}{A postive integer of number of days to achieve target}

\item{return_graph}{TRUE or FALSE on whether to return a graphical
visualization instead of a bmi value.}
}
\value{
A numeric number of body mass index value.
}
\description{
Compute average BMI change per week based on current weight, height, age,
and target BMI. The average BMI increase or decrease each week is computed.
If \code{return_graph}is \code{TRUE}, a \code{ggplot} graph is given based on forcasted
BMI trajectory for the target timeframe.
}
\examples{
project_bmi(weight=108, height=1.88, target_bmi = 28, number_of_days=60, return_graph=FALSE)

-0.3
}
