% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bmi.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Calculate the body mass index (BMI) based on current weight and height.}
\usage{
calculate_bmi(weight, height, return_graph = FALSE)
}
\arguments{
\item{weight}{A numeric number of body weight value.}

\item{height}{A numeric number of height value.}

\item{return_graph}{TRUE or FALSE on whether to return a graphical
visualization instead of a bmi value.}
}
\value{
A numeric number of body mass index value (rounded to two decimals).
}
\description{
Body Mass Index (BMI) is a calculated measure of weight relative to height,
defined as body weight in kilograms divided by height in meters squared
(Keys et al., 1972). If \code{return_graph} is \code{TRUE}, an
ggplot object is returned instead of a bmi value.
}
\examples{
calculate_bmi(100, 1.85, return_graph=FALSE)

29.22

}
